api_notif/requirements.txt
fastapi==0.92.0
orjson==3.8.7
python-logstash==0.4.8
uvicorn==0.20.0
pymongo==4.3.3
motor==3.1.1
gunicorn==20.1.0
python-dotenv==1.0.0